refactor(viewer): modify stringifyPoints location to Annotation utils
