+++ libs/insight-viewer/src/Annotation/hooks/useCircleAnnotation/index.ts
@@ -4,10 +4,10 @@ import { useOverlayContext } from '../../../contexts'
 import { calculateCircleArea } from '../../utils/calculateCircleArea'
 import { getCircleConnectingLine } from './utils/getCircleConnectingLine'
 import { getCircleEndPoint } from '../../utils/getCircleEndPoint'
-import { getCircleRadiusByCenter } from '../../../utils/common/getCircleRadius'
+import { stringifyPoints } from '../../utils/stringifyPoints'
 import { getCircleTextPosition } from './utils/getCircleTextPosition'
+import { getCircleRadiusByCenter } from '../../../utils/common/getCircleRadius'
 import { modifyConnectingLine } from '../../../utils/common/modifyConnectingLine'
-import { stringifyPoints } from '../../../utils/common/stringifyPoints'
 
 import { HALF_OF_RULER_TEXT_BOX } from '../../../const'
 import type { AreaAnnotation } from '../../types'
