+++ libs/insight-viewer/src/Annotation/hooks/useCreatingDrawableAnnotation/utils/getArrowPosition.spec.ts
@@ -1,4 +1,4 @@
-import { Point } from '../../types'
+import { Point } from '../../../../types'
 import { getArrowPosition } from './getArrowPosition'
 
 describe('getArrowPosition:', () => {
