+++ libs/insight-viewer/src/Annotation/hooks/useRulerMeasurement/index.ts
@@ -2,7 +2,7 @@ import useTextBox from '../useTextBox'
 import { useOverlayContext } from '../../../contexts'
 
 import { modifyConnectingLine } from '../../../utils/common/modifyConnectingLine'
-import { stringifyPoints } from '../../../utils/common/stringifyPoints'
+import { stringifyPoints } from '../../utils/stringifyPoints'
 import { getRulerTextPosition } from './utils/getRulerTextPosition'
 import { getRulerConnectingLine } from './utils/getRulerConnectingLine'
