+++ libs/insight-viewer/src/Annotation/utils/stringifyPoints.spec.ts
