+++ libs/insight-viewer/src/Annotation/utils/stringifyPoints.ts
